fastapi-backend/requirements.txt
uvicorn>=0.17.6
fastapi[all]
python-socketio
numpy
scipy
matplotlib
Pillow